# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Pharma API
  version: 1.0.0
  description: API multiempresa para operação fiscal, estoque, vendas e gestão do SaaS.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento
...
paths:
...
  /api/v1/fiscal/analises:
    get:
      tags: [Fiscal]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "200": { description: Histórico, evidências e revisões }
    post:
      tags: [Fiscal]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "201": { description: Solicitação de classificação registrada }
  /api/v1/fiscal/analises/{id}/revisao:
    put:
      tags: [Fiscal]
      security: [{ bearerAuth: [], companyHeader: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200": { description: Resultado e revisão humana registrados }
  /api/v1/fiscal/analises/{id}/evidencias:
    post:
      tags: [Fiscal]
      security: [{ bearerAuth: [], companyHeader: [] }]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201": { description: Fonte legal vinculada à análise }
...
  /api/v1/suporte/tickets:
    get:
      tags: [Suporte]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "200": { description: Tickets da empresa }
    post:
      tags: [Suporte]
      security: [{ bearerAuth: [], companyHeader: [] }]
      responses:
        "201": { description: Ticket criado }
  /api/v1/financeiro/assinaturas:
    get:
      tags: [Interno]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Assinaturas e últimas faturas }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/desenvolvimento/releases:
    get:
      tags: [Interno]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Releases, aprovações e clientes liberados }
    post:
      tags: [Interno]
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Release criada }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    companyHeader:
      type: apiKey
      in: header
      name: x-company-id
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    Forbidden:
      description: Perfil sem permissão para a operação